rename package UI to UIConfig
