+++ electron-app/src/renderer/repositories/UIConfig/browserRepository.ts
