+++ electron-app/src/renderer/repositories/UIConfig/electronRepository.ts
