+++ electron-app/src/renderer/repositories/UIConfig/index.ts
