+++ electron-app/src/renderer/repositories/UIConfig/registry.ts
