+++ electron-app/src/renderer/repositories/UIConfig/repository.ts
